@@ -1,28 +1,34 @@
 {
-    "name": "pod-point/countries",
-    "description": "A complete and partial list of countries with a Laravel service provider to load them into config",
+    "name": "pod-point/i18n",
+    "description": "A package of useful components for internationalisation in Laravel applications.",
     "type": "library",
     "require": {
-        "php": ">=5.5.9",
+        "php": ">=7.1",
         "illuminate/config": "5.*",
         "illuminate/support": "5.*",
         "illuminate/view": "5.*",
-        "league/iso3166": "^2.1"
+        "league/iso3166": "^2.1",
+        "guzzlehttp/guzzle": "^6.3"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8",
-        "satooshi/php-coveralls": "^1.0"
+        "satooshi/php-coveralls": "^1.0",
+        "phpunit/phpunit": "^7.5"
     },
     "license": "MIT",
     "authors": [
         {
-            "name": "Edward Coleridge Smith",
-            "email": "edward.smith@pod-point.com"
+            "name": "Pod Point Software Team",
+            "email": "software@pod-point.com"
         }
     ],
     "autoload": {
         "psr-4": {
-            "PodPoint\\Countries\\": "src/"
+            "PodPoint\\I18n\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "PodPoint\\I18n\\Tests\\": "tests/"
         }
     }
 }
